adds suggested changes
* adds more doc comment changes
* moves properties like annoymous type counter and tera to generator
from context
* context only stores data model property and is locally created for
each ISL type code generation
* adds Template struct which stores information about which templates
are supported and has a method to navogate to a particular template file
for code generation

+++ src/bin/ion/commands/beta/generate/mod.rs
@@ -11,7 +11,7 @@ use clap::{Arg, ArgAction, ArgMatches, Command};
 use ion_schema::authority::{DocumentAuthority, FileSystemDocumentAuthority};
 use ion_schema::system::SchemaSystem;
 use std::fs;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 
 pub struct GenerateCommand;
 
@@ -35,20 +35,21 @@ impl IonCliCommand for GenerateCommand {
             .arg(
                 Arg::new("schema")
                     .long("schema")
+                    .required(true)
                     .short('s')
-                    .help("Schema file [default: no schema will be applied]"),
+                    .help("Schema file"),
             )
             .arg(
                 Arg::new("language")
                     .long("language")
                     .short('l')
-                    .default_value("rust")
+                    .required(true)
                     .value_parser(["java", "rust"])
                     .help("Programming language for the generated code"),
             )
             .arg(
                 // Directory(s) that will be used as authority(s) for schema system
-                Arg::new("directories")
+                Arg::new("directory")
                     .long("directory")
                     .short('d')
                     // If this appears more than once, collect all values
@@ -64,10 +65,16 @@ impl IonCliCommand for GenerateCommand {
         let language: Language = args.get_one::<String>("language").unwrap().as_str().into();
 
         // Extract output path information where the generated code will be saved
-        let output = Path::new(args.get_one::<String>("output").unwrap());
+        // Create a module `ion_data_model` for storing all the generated code in the output directory
+        let binding = match args.get_one::<String>("output") {
+            Some(output_path) => PathBuf::from(output_path).join("ion_data_model"),
+            None => PathBuf::from("./ion_data_model"),
+        };
+
+        let output = binding.as_path();
 
         // Extract the user provided document authorities/ directories
-        let authorities: Vec<&String> = args.get_many("directories").unwrap().collect();
+        let authorities: Vec<&String> = args.get_many("directory").unwrap().collect();
 
         // Extract schema file provided by user
         let schema_id = args.get_one::<String>("schema").unwrap();
@@ -86,13 +93,21 @@ impl IonCliCommand for GenerateCommand {
 
         let schema = schema_system.load_isl_schema(schema_id).unwrap();
 
-        // clean the target output directory before generating new code
-        fs::remove_dir_all(output).unwrap();
-        fs::create_dir(output).unwrap();
+        // clean the target output directory if it already exists, before generating new code
+        if output.exists() {
+            fs::remove_dir_all(output).unwrap();
+        }
+        fs::create_dir_all(output).unwrap();
+
+        println!("Started generating code...");
 
         // generate code based on schema and programming language
-        CodeGenerator::generate_code(language, schema, output).unwrap();
+        CodeGenerator::new(language, output)
+            .generate(schema)
+            .unwrap();
 
+        println!("Code generation complete successfully!");
+        println!("Path to generated code: {}", output.display());
         Ok(())
     }
 }
